Extract ProjectHandler from CalciteRelNodeVisitor
